requirements.txt
aioconsole==0.7.1
colored==2.2.4
openai==1.30.1
python-dotenv==1.0.1
